requirements.txt
httpx>=0.27
openai>=1.30
google-genai>=1.0.0
python-dotenv>=1.0